# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Bridge API
  version: 1.125.0
  description: 'Bridge — federated, self-hosted communication platform REST API.


    **Authentication:** Tüm korumalı endpoint''ler `Authorization: Bearer <accessToken>` header''ı gerektirir.

    Access token''lar kısa ömürlüdür (15 dk). `POST /api/refresh` ile yenilenebilir (httpOnly cookie).


    **Rate Limiting:** Endpoint başına granüler rate limit uygulanır. 429 döndüğünde `Retry-After` header''ını kontrol edin.


    **Federation:** Bridge sunucular arası ActivityPub protokolü kullanır. Federation endpoint''leri `/api/federation` altındadır.

    '
  contact:
    name: Bridge Contributors
    url: https://github.com/bridge-app/bridge
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: https://your-bridge-instance.com/api
  description: Production
- url: http://localhost:3001/api
  description: Local development
...
paths:
...
  /discover:
    get:
      tags:
      - Discover
      summary: Sunucu keşif listesi
      parameters:
      - in: query
        name: category
        schema:
          type: string
          enum:
          - gaming
          - music
          - art
          - tech
          - edu
          - social
          - other
      - in: query
        name: q
        schema:
          type: string
        description: Arama sorgusu
      responses:
        200:
          description: Keşif listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
      operationId: get_discover
...
  /discover/settings:
    patch:
      tags:
      - Discover
      summary: Sunucu keşif ayarları güncelle
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                serverId:
                  type: string
                category:
                  type: string
                listed:
                  type: boolean
      responses:
        200:
          description: Ayarlar güncellendi
      operationId: patch_discover_settings
  /discover/admin/feature:
    post:
      tags:
      - Discover
      - Admin
      summary: Sunucuyu öne çıkar / kaldır
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - serverId
              - featured
              properties:
                serverId:
                  type: string
                featured:
                  type: boolean
      responses:
        200:
          description: Öne çıkarma durumu güncellendi
        403:
          $ref: '#/components/responses/Forbidden'
      operationId: post_discover_admin_feature
...
  /federation/discover:
    get:
      tags:
      - Federation
      summary: Yeni peer'leri otomatik kesif
      security:
      - bearerAuth: []
      responses:
        200:
          description: Bulunan peer listesi
      operationId: get_federation_discover
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. POST /api/login ile alınır, süresi 15 dakikadır.
    botToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bot token: `Authorization: Bot brg_bot_…` (sunucu sahibi bot oluştururken bir kez gösterilir).'
...
  responses:
...
    Forbidden:
      description: Yetersiz izin
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Server:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        iconUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        ownerId:
          type: string
        isPublic:
          type: boolean
        isVerified:
          type: boolean
        memberCount:
          type: integer
        category:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        createdAt:
          type: integer
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
    Channel:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        serverId:
          type: string
        type:
          type: string
          enum:
          - text
          - voice
          - video
          - announcement
          - stage
          - forum
        position:
          type: integer
        topic:
          type: string
          nullable: true
        nsfw:
          type: boolean
        slowMode:
          type: integer
          description: Saniye cinsinden yavaş mod süresi
        parentId:
          type: string
          nullable: true
          description: Kategori kanal ID'si
        e2eeEnabled:
          type: boolean
        createdAt:
          type: integer
...
    Role:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        position:
          type: integer
        permissions:
          type: integer
          description: Bitfield izin değeri
        hoist:
          type: boolean
        mentionable:
          type: boolean
...
security:
- bearerAuth: []